+++ playlet-web/package.json
@@ -23,6 +23,7 @@
     "tailwindcss": "^3.3.2",
     "tslib": "^2.5.3",
     "typescript": "^5.1.3",
-    "vite": "^4.3.9"
+    "vite": "^4.3.9",
+    "vite-tsconfig-paths": "^4.2.0"
   }
 }
